+++ CHANGELOG.md
@@ -1,7 +1,12 @@
 # Unreleased
 
-Changed
+Added
+  * AZ-410: Added a variable to enable choosing a TLS version for the logs storage account (Defaulted to TLS1_2)
+
+Updated
   * AZ-325: Upgrade keyvault module to latest version + Update CI
+
+Fixed
   * AZ-326: Add missing keyvault's inputs
 
 # v3.1.2/v4.0.0 - 2020-12-11